requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
fonttools==4.39.4
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
djangorestframework-simplejwt==5.1.0
